+++ src/main/java/qodous/erp/inventory/security/services/IRoleService.java
@@ -5,5 +5,5 @@ import qodous.erp.inventory.security.domain.Role;
 import java.util.Optional;
 
 public interface IRoleService {
-    public Optional<Role> findById(int roleId);
+    public Optional<Role> findById(Long roleId);
 }
